+++ CMakeLists.txt
@@ -23,6 +23,10 @@ if (DEBUG)
   add_definitions(-DDEBUG=1)
 endif()
 
+if (USE_DPUCZDX8G_VART)
+  add_definitions(-DUSE_DPUCZDX8G_VART)
+endif()
+
 if (USE_VAI_RT_DPUCADX8G)
   set(vitisai_ROOT_DIR $ENV{CONDA_PREFIX})
   set(vitisai_INCLUDE_DIRS ${vitisai_ROOT_DIR}/lib/python3.6/site-packages/vai/dpuv1/rt/vitis/include/dpu)
@@ -41,6 +45,24 @@ if (USE_VAI_RT_DPUCADX8G)
   file(GLOB_RECURSE vitisai_default_SOURCES "src/pyxir/runtime/backends/vai_rt/xir_api/*.cpp")
   list(REMOVE_ITEM SOURCES ${vitisai_default_SOURCES})
 
+elseif (USE_VAI_RT_DPUCAHX8H OR (USE_VAI_RT_DPUCZDX8G AND USE_DPUCZDX8G_VART))
+  set(vitisai_LIBRARIES /usr/lib)
+  set(VITISAI_SHARED_LIB libvart-runner.so)  
+  add_library(vitisai_rt SHARED IMPORTED)
+  set_property(TARGET vitisai_rt PROPERTY IMPORTED_LOCATION ${vitisai_LIBRARIES}/${VITISAI_SHARED_LIB})
+  list(APPEND pyxir_EXT_LIBRARIES vitisai_rt)
+  if (USE_VAI_RT_DPUCAHX8H)
+    add_definitions(-DUSE_VAI_RT_DPUCAHX8H=1)
+  else()
+    add_definitions(-DUSE_VAI_RT_DPUCZDX8G=1)
+  endif()
+
+  # Remove VAI_DEFAULT backend source files
+  file(GLOB_RECURSE vitisai_default_SOURCES "src/pyxir/runtime/backends/vai_default/*.cpp")
+  list(REMOVE_ITEM SOURCES ${vitisai_default_SOURCES})
+  file(GLOB_RECURSE vitisai_default_SOURCES "src/pyxir/runtime/backends/vai_rt/vai_api/*.cpp")
+  list(REMOVE_ITEM SOURCES ${vitisai_default_SOURCES})
+
 elseif (USE_VAI_RT_DPUCZDX8G)
   set(vitisai_INCLUDE_DIRS /usr/include/vai)
   set(vitisai_LIBRARIES /usr/lib)
@@ -55,20 +77,6 @@ elseif (USE_VAI_RT_DPUCZDX8G)
   file(GLOB_RECURSE vitisai_default_SOURCES "src/pyxir/runtime/backends/vai_default/*.cpp")
   list(REMOVE_ITEM SOURCES ${vitisai_default_SOURCES})
   file(GLOB_RECURSE vitisai_default_SOURCES "src/pyxir/runtime/backends/vai_rt/xir_api/*.cpp")
-  list(REMOVE_ITEM SOURCES ${vitisai_default_SOURCES})
-
-elseif (USE_VAI_RT_DPUCAHX8H)
-  set(vitisai_LIBRARIES /usr/lib)
-  set(VITISAI_SHARED_LIB libvart-runner.so)  
-  add_library(vitisai_rt SHARED IMPORTED)
-  set_property(TARGET vitisai_rt PROPERTY IMPORTED_LOCATION ${vitisai_LIBRARIES}/${VITISAI_SHARED_LIB})
-  list(APPEND pyxir_EXT_LIBRARIES vitisai_rt)
-  add_definitions(-DUSE_VAI_RT_DPUCAHX8H=1)
-
-  # Remove VAI_DEFAULT backend source files
-  file(GLOB_RECURSE vitisai_default_SOURCES "src/pyxir/runtime/backends/vai_default/*.cpp")
-  list(REMOVE_ITEM SOURCES ${vitisai_default_SOURCES})
-  file(GLOB_RECURSE vitisai_default_SOURCES "src/pyxir/runtime/backends/vai_rt/vai_api/*.cpp")
   list(REMOVE_ITEM SOURCES ${vitisai_default_SOURCES})
 
 else()
