+++ src/pyxir/runtime/backends/vai_rt/vai_compute_func.hpp
@@ -19,10 +19,10 @@
 #include <memory>
 #include <unordered_set>
 
-#if defined(USE_VAI_RT_DPUCADX8G) || defined(USE_VAI_RT_DPUCZDX8G)
+#if defined(USE_VAI_RT_DPUCADX8G) || (defined(USE_VAI_RT_DPUCZDX8G) && !defined(USE_DPUCZDX8G_VART))
 #include <dpu_runner.hpp>
 #include "vai_api/dpu_func.hpp"
-#elif defined(USE_VAI_RT_DPUCAHX8H)
+#elif defined(USE_VAI_RT_DPUCAHX8H) || (defined(USE_VAI_RT_DPUCZDX8G) && defined(USE_DPUCZDX8G_VART))
 #include "xir_api/dpu_func.hpp"
 #endif
