+++ tests/unit/contrib/compilation_infra.py
@@ -20,11 +20,12 @@ import shutil
 import numpy as np
 import pyxir as px
 
+from pyxir.target_registry import TargetRegistry
 from pyxir.graph import XGraph
 from pyxir.graph.xgraph_factory import XGraphFactory
-from pyxir.quantization.decent_quantizer import DECENTQuantizer
 
 XGRAPH_FACTORY = XGraphFactory()
+TARGET_REGISTRY = TargetRegistry()
 FILE_PATH = os.path.dirname(os.path.realpath(__file__))
 
 
@@ -112,44 +113,47 @@ def conv2d_pool2d_nhwc_oihw_test(
     conv_groups=1,
     conv_invalid=False,
     kernel_layout="OIHW",
-    target="DPUCZDX8G-zcu104",
+    targets=["DPUCZDX8G-zcu104"],
 ) -> None:
 
-    xgraph = _create_conv2d_pool2d_nhwc_oihw(
-        in_shape,
-        w_shape,
-        conv_padding,
-        conv_strides,
-        conv_dilation,
-        pool_type,
-        pool_size,
-        pool_padding,
-        pool_strides,
-        conv_groups,
-        conv_invalid,
-        kernel_layout,
-        target,
-    )
+    for target in targets:
+        xgraph = _create_conv2d_pool2d_nhwc_oihw(
+            in_shape,
+            w_shape,
+            conv_padding,
+            conv_strides,
+            conv_dilation,
+            pool_type,
+            pool_size,
+            pool_padding,
+            pool_strides,
+            conv_groups,
+            conv_invalid,
+            kernel_layout,
+            target,
+        )
 
-    def inputs_func(iter):
-        inputs = np.ones(in_shape, dtype=np.float32)
-        return {"in1": inputs}
+        def inputs_func(iter):
+            inputs = np.ones(in_shape, dtype=np.float32)
+            return {"in1": inputs}
 
-    work_dir = os.path.join(FILE_PATH, "work")
-    build_dir = os.path.join(FILE_PATH, "build")
-    quantizer = DECENTQuantizer(xgraph, inputs_func, work_dir=work_dir)
-    q_xgraph = quantizer.quantize()
-    opt_xgraph = px.optimize(q_xgraph, target)
-    c_xgraph = px.compile(opt_xgraph, target, work_dir=work_dir, build_dir=build_dir)
-    c_output = c_xgraph.get_compiler_output()
+        work_dir = os.path.join(FILE_PATH, "work")
+        build_dir = os.path.join(FILE_PATH, "build")
+        quantize_func = TARGET_REGISTRY.get_target_quantizer(target)
+        q_xgraph = quantize_func(xgraph, inputs_func, work_dir=work_dir)
+        opt_xgraph = px.optimize(q_xgraph, target)
+        c_xgraph = px.compile(
+            opt_xgraph, target, work_dir=work_dir, build_dir=build_dir
+        )
+        c_output = c_xgraph.get_compiler_output()
 
-    assert list(c_output.keys()) == ["xp0"]
-    assert c_output.get_in_map("xp0") == {"xinput0": "xinput0:0"}
-    assert c_output.get_out_map("xp0") == {"pool1": "pool1:0"}
-    assert len(c_output.get_code_files("xp0")) == 1
+        assert list(c_output.keys()) == ["xp0"]
+        assert c_output.get_in_map("xp0") == {"xinput0": "xinput0:0"}
+        assert c_output.get_out_map("xp0") == {"pool1": "pool1:0"}
+        assert len(c_output.get_code_files("xp0")) == 1
 
-    shutil.rmtree(work_dir)
-    shutil.rmtree(build_dir)
+        shutil.rmtree(work_dir)
+        shutil.rmtree(build_dir)
 
 
 def xcompiler_conv2d_pool2d_nhwc_oihw_test(
@@ -165,53 +169,56 @@ def xcompiler_conv2d_pool2d_nhwc_oihw_test(
     conv_groups=1,
     conv_invalid=False,
     kernel_layout="OIHW",
-    target="DPUCAHX8H-u50",
+    targets=["DPUCAHX8H-u50"],
     expected_nb_subgraphs=3,
 ):
 
-    xgraph = _create_conv2d_pool2d_nhwc_oihw(
-        in_shape,
-        w_shape,
-        conv_padding,
-        conv_strides,
-        conv_dilation,
-        pool_type,
-        pool_size,
-        pool_padding,
-        pool_strides,
-        conv_groups,
-        conv_invalid,
-        kernel_layout,
-        target,
-    )
+    for target in targets:
+        xgraph = _create_conv2d_pool2d_nhwc_oihw(
+            in_shape,
+            w_shape,
+            conv_padding,
+            conv_strides,
+            conv_dilation,
+            pool_type,
+            pool_size,
+            pool_padding,
+            pool_strides,
+            conv_groups,
+            conv_invalid,
+            kernel_layout,
+            target,
+        )
 
-    def inputs_func(iter):
-        inputs = np.ones(in_shape, dtype=np.float32)
-        return {"in1": inputs}
+        def inputs_func(iter):
+            inputs = np.ones(in_shape, dtype=np.float32)
+            return {"in1": inputs}
 
-    work_dir = os.path.join(FILE_PATH, "work")
-    build_dir = os.path.join(FILE_PATH, "build")
-    quantizer = DECENTQuantizer(xgraph, inputs_func, work_dir=work_dir)
-    q_xgraph = quantizer.quantize()
-    opt_xgraph = px.optimize(q_xgraph, target)
-    c_xgraph = px.compile(opt_xgraph, target, work_dir=work_dir, build_dir=build_dir)
-    c_output = c_xgraph.get_compiler_output()
+        work_dir = os.path.join(FILE_PATH, "work")
+        build_dir = os.path.join(FILE_PATH, "build")
+        quantize_func = TARGET_REGISTRY.get_target_quantizer(target)
+        q_xgraph = quantize_func(xgraph, inputs_func, work_dir=work_dir)
+        opt_xgraph = px.optimize(q_xgraph, target)
+        c_xgraph = px.compile(
+            opt_xgraph, target, work_dir=work_dir, build_dir=build_dir
+        )
+        c_output = c_xgraph.get_compiler_output()
 
-    assert list(c_output.keys()) == ["xp0"]
-    assert c_output.get_in_map("xp0") == {"xinput0": "xinput0"}
-    assert c_output.get_out_map("xp0") == {"pool1": "pool1"}
-    assert len(c_output.get_code_files("xp0")) == 1
+        assert list(c_output.keys()) == ["xp0"]
+        assert c_output.get_in_map("xp0") == {"xinput0": "xinput0"}
+        assert c_output.get_out_map("xp0") == {"pool1": "pool1"}
+        assert len(c_output.get_code_files("xp0")) == 1
 
-    g = xir.Graph.deserialize(os.path.join(build_dir, "xp0.xmodel"))
-    # TODO subgraphs[1].get_attr("device") -> *** RuntimeError: bad any_cast
-    subgraphs = get_child_subgraphs(g)
-    assert len(subgraphs) == expected_nb_subgraphs
-    dpu_subgraph = subgraphs[1]
-    # import pdb; pdb.set_trace()
-    # assert len(dpu_subgraph.get_children()) == 3
+        g = xir.Graph.deserialize(os.path.join(build_dir, "xp0.xmodel"))
+        # TODO subgraphs[1].get_attr("device") -> *** RuntimeError: bad any_cast
+        subgraphs = get_child_subgraphs(g)
+        assert len(subgraphs) == expected_nb_subgraphs
+        dpu_subgraph = subgraphs[1]
+        # import pdb; pdb.set_trace()
+        # assert len(dpu_subgraph.get_children()) == 3
 
-    shutil.rmtree(work_dir)
-    shutil.rmtree(build_dir)
+        shutil.rmtree(work_dir)
+        shutil.rmtree(build_dir)
 
 
 def _create_scale_conv2d_nhwc_oihw(
@@ -310,8 +317,8 @@ def xcompiler_scale_conv2d_nhwc_oihw_test(
 
     work_dir = os.path.join(FILE_PATH, "work")
     build_dir = os.path.join(FILE_PATH, "build")
-    quantizer = DECENTQuantizer(xgraph, inputs_func, work_dir=work_dir)
-    q_xgraph = quantizer.quantize()
+    quantize_func = TARGET_REGISTRY.get_target_quantizer(target)
+    q_xgraph = quantize_func(xgraph, inputs_func, work_dir=work_dir)
     opt_xgraph = px.optimize(q_xgraph, target)
     c_xgraph = px.compile(opt_xgraph, target, work_dir=work_dir, build_dir=build_dir)
     c_output = c_xgraph.get_compiler_output()
@@ -483,8 +490,8 @@ def xcompiler_resnetv1_block_test(
 
     work_dir = os.path.join(FILE_PATH, "work")
     build_dir = os.path.join(FILE_PATH, "build")
-    quantizer = DECENTQuantizer(xgraph, inputs_func, work_dir=work_dir)
-    q_xgraph = quantizer.quantize()
+    quantize_func = TARGET_REGISTRY.get_target_quantizer(target)
+    q_xgraph = quantize_func(xgraph, inputs_func, work_dir=work_dir)
     opt_xgraph = px.optimize(q_xgraph, target)
     c_xgraph = px.compile(opt_xgraph, target, work_dir=work_dir, build_dir=build_dir)
     c_output = c_xgraph.get_compiler_output()
@@ -500,3 +507,125 @@ def xcompiler_resnetv1_block_test(
 
     shutil.rmtree(work_dir)
     shutil.rmtree(build_dir)
+
+
+def _create_conv2d_leaky_relu_nhwc_oihw(
+    in_shape,
+    w_shape,
+    conv_padding,
+    conv_strides,
+    conv_dilation,
+    kernel_layout="OIHW",
+    target="DPUCZDX8G-zcu104",
+) -> XGraph:
+
+    kernel_w, kernel_h = w_shape[2], w_shape[3]
+    W = np.random.randint(-10, 10, size=w_shape).astype(np.float32)
+    # B = np.array([1., -1.], dtype=np.float32)
+
+    x1 = px.ops.input("in1", shape=list(in_shape))
+    w1 = px.ops.constant("weight", W)
+    conv1 = px.ops.conv2d(
+        op_name="conv1",
+        input_layer=x1,
+        weights_layer=w1,
+        kernel_size=[kernel_w, kernel_h],
+        strides=list(conv_strides),
+        padding_hw=list(conv_padding),
+        dilation=list(conv_dilation),
+        data_layout="NHWC",
+    )
+    lr1 = px.ops.leaky_relu("lr1", [conv1], alpha=0.1)
+    net = [x1, conv1, lr1]
+    xgraph = XGRAPH_FACTORY.build_from_xlayer(net)
+    xgraph = px.partition(xgraph, [target])
+    return xgraph
+
+
+def conv2d_leaky_relu_nhwc_oihw_test(
+    in_shape,
+    w_shape,
+    conv_padding,
+    conv_strides,
+    conv_dilation,
+    kernel_layout="OIHW",
+    targets=["DPUCZDX8G-zcu104"],
+) -> None:
+
+    for target in targets:
+        xgraph = _create_conv2d_leaky_relu_nhwc_oihw(
+            in_shape,
+            w_shape,
+            conv_padding,
+            conv_strides,
+            conv_dilation,
+            kernel_layout,
+            target,
+        )
+
+        def inputs_func(iter):
+            inputs = np.ones(in_shape, dtype=np.float32)
+            return {"in1": inputs}
+
+        work_dir = os.path.join(FILE_PATH, "work")
+        build_dir = os.path.join(FILE_PATH, "build")
+        quantize_func = TARGET_REGISTRY.get_target_quantizer(target)
+        q_xgraph = quantize_func(xgraph, inputs_func, work_dir=work_dir)
+        opt_xgraph = px.optimize(q_xgraph, target)
+        c_xgraph = px.compile(
+            opt_xgraph, target, work_dir=work_dir, build_dir=build_dir
+        )
+        c_output = c_xgraph.get_compiler_output()
+
+        assert list(c_output.keys()) == ["xp0"]
+        assert c_output.get_in_map("xp0") == {"xinput0": "xinput0:0"}
+        assert c_output.get_out_map("xp0") == {"lr1": "lr1:0"}
+        assert len(c_output.get_code_files("xp0")) == 1
+
+        shutil.rmtree(work_dir)
+        shutil.rmtree(build_dir)
+
+
+def xcompiler_conv2d_leaky_relu_nhwc_oihw_test(
+    in_shape,
+    w_shape,
+    conv_padding,
+    conv_strides,
+    conv_dilation,
+    kernel_layout="OIHW",
+    targets=["DPUCZDX8G-zcu104"],
+    expected_nb_subgraphs=3,
+) -> None:
+
+    for target in targets:
+        xgraph = _create_conv2d_leaky_relu_nhwc_oihw(
+            in_shape,
+            w_shape,
+            conv_padding,
+            conv_strides,
+            conv_dilation,
+            kernel_layout,
+            target,
+        )
+
+        def inputs_func(iter):
+            inputs = np.ones(in_shape, dtype=np.float32)
+            return {"in1": inputs}
+
+        work_dir = os.path.join(FILE_PATH, "work")
+        build_dir = os.path.join(FILE_PATH, "build")
+        quantize_func = TARGET_REGISTRY.get_target_quantizer(target)
+        q_xgraph = quantize_func(xgraph, inputs_func, work_dir=work_dir)
+        opt_xgraph = px.optimize(q_xgraph, target)
+        c_xgraph = px.compile(
+            opt_xgraph, target, work_dir=work_dir, build_dir=build_dir
+        )
+        
+        g = xir.Graph.deserialize(os.path.join(build_dir, "xp0.xmodel"))
+        # TODO subgraphs[1].get_attr("device") -> *** RuntimeError: bad any_cast
+        subgraphs = get_child_subgraphs(g)
+        assert (
+            len(subgraphs) == expected_nb_subgraphs
+        ), "Expected {0} subgraphs but got: {1}".format(
+            expected_nb_subgraphs, len(subgraphs)
+        )
